Add files via upload
Generative Adversarial Networks (GANs) are a type of neural network that can generate new data samples that are similar to the training data. In this project, we  create and train a small GAN from scratch on the MNIST dataset. We  display a few generated images as training progresses and inspect them visually. The goal is to generate new images that are similar to the ones in the dataset We will also implement the idea of ’Inception scores’ to evaluate our model.

Overall, these results suggest that the GAN model is able to generate high-quality images that closely resembled the training data, and that the model continued to improve as it was trained for more epochs.

To improve our result in the future, we can use a more big dataset because The model was trained on a relatively small dataset (60,000 images), which may limit its ability to generalize to new data. The generated images were limited to grayscale and did not capture the full range of colors and textures present in the training data. The model was trained for a fixed number of epochs and did not incorporate any early stopping or other techniques to prevent overfitting so we can also add the number of epochs.